+++ lib/src/widgets/pdf_page_text_overlay.dart
@@ -4,6 +4,8 @@ import 'package:flutter/rendering.dart';
 import '../../pdfrx.dart';
 
 /// A widget that displays selectable text on a page.
+///
+/// If [PdfDocument.permissions] does not allow copying, the widget does not show anything.
 class PdfPageTextOverlay extends StatefulWidget {
   const PdfPageTextOverlay({
     required this.registrar,
@@ -46,7 +48,10 @@ class _PdfPageTextOverlayState extends State<PdfPageTextOverlay> {
 
   @override
   Widget build(BuildContext context) {
-    if (pageText == null) return const SizedBox();
+    if (pageText == null ||
+        widget.page.document.permissions?.allowsCopying == false) {
+      return const SizedBox();
+    }
     return Positioned(
       left: widget.pageRect.left,
       top: widget.pageRect.top,